+++ ncp/stores/__init__.py
@@ -1,7 +1,8 @@
 """Store backends package."""
 
+from ncp.stores.factory import create_store
 from ncp.stores.pgvector import PgvectorStore
 from ncp.stores.redis import RedisStore
 from ncp.stores.sqlite import SQLiteStore
 
-__all__ = ["SQLiteStore", "RedisStore", "PgvectorStore"]
+__all__ = ["SQLiteStore", "RedisStore", "PgvectorStore", "create_store"]